packs/vdx/etc/opt/vdx_pack_gen/requirements.txt
docutils
pyaml
jinja2
